feat: Enhance A9G GPS Tracker documentation and refactor HTTP handling
- Updated README.md to include comprehensive overview, features, hardware requirements, and configuration details for the A9G GPS Tracker.
- Added sections for getting started, status indicators, and advanced features like Traccar server integration and APN re-activation workaround.
- Refactored HTTP handling in gps_tracker.c to simplify the HTTP/HTTPS POST request process by consolidating SSL configuration and removing redundant code.
- Introduced a new http.c and http.h file to encapsulate HTTP-related functions, improving code organization and maintainability.
- Removed obsolete APN re-activation workaround logic from network.c and centralized it in a dedicated function.
- Updated network.h to declare new APN handling functions and removed unnecessary SSL configuration from the global scope.
