+++ app/src/gps_tracker.h
@@ -16,8 +16,6 @@
 #define DEFAULT_LOG_LEVEL         "info"
 #define DEFAULT_LOG_OUTPUT        "uart"
 
-void apnWorkaround_init(void);
-bool gsm_AttachActivate(void);
 void gps_trackerTask(void *pData);
 
 #endif
