+++ app/src/network.h
@@ -4,28 +4,6 @@
 #define SSL_WRITE_TIMEOUT 5000
 #define SSL_READ_TIMEOUT  2000
 
-/**
- * @brief Sends an HTTP/HTTPs POST request to the specified server.
- * 
- * @param hostName The hostName name of the server (e.g., "example.com").
- * @param port The port number to connect to (e.g., 443 for HTTPS).
- * @param path The path of the resource on the server (e.g., "/api/data").
- * @param data The data to be sent in the POST request body.
- * @param dataLen The length of the data to be sent.
- * @param retBuffer A buffer to store the response data received from the server.
- * @param retBufferSize The size of the buffer allocated for the response data.
- * @return int Returns the length of the response on success, or -1 on failure.
- */
-int Http_Post(SSL_Config_t *sslConfig,
-              const char   *hostName,
-              const char   *port,
-              const char   *path,
-              const char   *data,
-              uint16_t      dataLen,
-              char*         retBuffer,
-              int           retBufferSize);
-
-
 /*
  * @brief Global variable to store cell information.
  *
@@ -46,5 +24,24 @@ void networkCellInfoTimer(HANDLE taskHandle);
  * @param number The number of cells to process (used for validation).
  */
 void networkCellInfoCallback(Network_Location_t* loc, int number);
+
+/**
+ * @brief Initializes the APN re-activation workaround.
+ *
+ * This function sets up the dummy APN context and initializes the workaround state.
+ * It should be called once during system initialization. 
+ */
+void apnWorkaround_init(void);
+
+/**
+ * @brief Attach and activate the GSM network.
+ *
+ * This function checks the current attach and activate status of the GSM network.
+ * If not attached, it attempts to attach. If attached but not activated, it activates the network.
+ * It handles the APN re-activation workaround by toggling between a dummy APN and the real APN.
+ * @return true if the operation was successful, false otherwise.
+ */
+bool gsm_AttachActivate(void);
+
 
 #endif
